feat(supplyOrder): implement create SupplyOrder and SupplyOrderItem functionality
- Add handlers, routers, DTOs, storage, and services for SupplyOrder and SupplyOrderItem
- Ensure relationships between SupplyOrder and SupplyOrderItem are correctly implemented
- Update existing handlers and repositories to accommodate new SupplyOrder features

+++ internal/entity/dtos/sql/ingredient/create_ingredient.go
@@ -3,7 +3,7 @@ package ingredientmodel
 import "tart-shop-manager/internal/common"
 
 type CreateIngredient struct {
-	IngredientID uint   `gorm:"column:ingredient_id;primaryKey;autoIncrement" json:"-"`
+	IngredientID uint64 `gorm:"column:ingredient_id;primaryKey;autoIncrement" json:"-"`
 	Name         string `gorm:"column:name;size:200;not null" json:"name"`
 	Description  string `gorm:"column:description;type:text" json:"description"`
 	Unit         string `gorm:"column:unit;size:100;not null" json:"unit"`
